+++ src/scripts/constants/constants.js
@@ -1,0 +1,7 @@
+const CONSTANTS = {
+  MODULE_ID: "magic-items-2",
+  PATH: `modules/magic-items-2/`,
+  FLAGS: {},
+};
+CONSTANTS.PATH = `modules/${CONSTANTS.MODULE_NAME}/`;
+export default CONSTANTS;
